+++ src/components/Skeleton/SearchSkeleton/index.jsx
